backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pydantic>=2.6.0
google-generativeai>=0.5.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
transformers>=4.35.0
torch>=2.1.0
moviepy>=1.0.3
pillow>=10.0.0
sqlalchemy>=2.0.0
pymysql>=1.1.0
cryptography>=41.0.0
alembic>=1.13.0
requests>=2.31.0


